主体最终版

+++ php/modify.php
@@ -10,54 +10,55 @@ include_once "connectDatabase.php";
 $uid=$_POST['uid'];
 $title=$_POST['photoName'];
 $message=$_POST['description'];
-$path=$_FILES['file']['name'];
 $imageID=$_POST['ImageID'];
 // 允许上传的图片后缀
 $allowedExts = array("gif", "jpeg", "jpg", "png");
-$temp = explode(".", $_FILES["file"]["name"]);
-$extension = end($temp);     // 获取文件后缀名
-if ((($_FILES["file"]["type"] == "image/gif")
-        || ($_FILES["file"]["type"] == "image/jpeg")
-        || ($_FILES["file"]["type"] == "image/jpg")
-        || ($_FILES["file"]["type"] == "image/pjpeg")
-        || ($_FILES["file"]["type"] == "image/x-png")
-        || ($_FILES["file"]["type"] == "image/png"))
-    && ($_FILES["file"]["size"] < 10000*1024)   // 小于 5000 kb
-)
-{
-    if ($_FILES["file"]["error"] > 0)
-    {
-        echo "错误：: " . $_FILES["file"]["error"] . "<br>";
-    }
-    else
-    {
+if($_FILES["file"]["size"]!=0) {
+    $path=$_FILES['file']['name'];
+    $temp = explode(".", $_FILES["file"]["name"]);
+    $extension = end($temp);     // 获取文件后缀名
+    if ((($_FILES["file"]["type"] == "image/gif")
+            || ($_FILES["file"]["type"] == "image/jpeg")
+            || ($_FILES["file"]["type"] == "image/jpg")
+            || ($_FILES["file"]["type"] == "image/pjpeg")
+            || ($_FILES["file"]["type"] == "image/x-png")
+            || ($_FILES["file"]["type"] == "image/png"))
+        && ($_FILES["file"]["size"] < 10000 * 1024)   // 小于 10000 kb
+    ) {
+        if ($_FILES["file"]["error"] > 0) {
+            die("错误：: " . $_FILES["file"]["error"] . "<br>");
+        } else {
 
-        // 判断../img/travel-images/large/目录是否存在该文件
-        // 如果没有../img/travel-images/large/目录，你需要创建它
-        while (file_exists("../img/travel-images/large/" . $path))
-        {
-            echo $path . " 文件已经存在。 ";
-            $path=$path+rand(1000, 9999);
+            // 判断../img/travel-images/large/目录是否存在该文件
+            // 如果没有../img/travel-images/large/目录，你需要创建它
+            while (file_exists("../img/travel-images/large/" . $path)) {
+                $path = $path + rand(1000, 9999);
+            }
+
+            // 如果 upload 目录不存在该文件则将文件上传到 upload 目录下
+            move_uploaded_file($_FILES["file"]["tmp_name"], "../img/travel-images/large/" . $path);
+            clipPic();
+            $db->begin_transaction();
+            changeImage();
+            changePost();
+            changeModifyTime();
+            $db->commit();
+            echo "Done";
+
         }
-
-        // 如果 upload 目录不存在该文件则将文件上传到 upload 目录下
-        move_uploaded_file($_FILES["file"]["tmp_name"], "../img/travel-images/large/" . $path);
-        clipPic();
-        $db->begin_transaction();
-        changeImage();
-        changePost();
-        changeModifyTime();
-        $db->commit();
-        echo "Done";
-
+    } else {
+        echo "You should upload an image file of .jpg, .jpeg, .png, .gif";
+        var_dump($_FILES);
+        var_dump($_POST);
+        echo $_FILES['file']['error'];
     }
-}
-else
-{
-    echo "You should upload an image file of .jpg, .jpeg, .png, .gif";
-    var_dump($_FILES);
-    var_dump($_POST);
-    echo $_FILES['file']['error'];
+}else{
+    $db->begin_transaction();
+    changeImage();
+    changePost();
+    changeModifyTime();
+    $db->commit();
+    echo "Done";
 }
 
 function changeModifyTime(){
@@ -68,23 +69,25 @@ function changeModifyTime(){
 }
 
 function changeImage(){
-    global $db, $imageID, $path;
+    global $db, $imageID;
 
-    $rsOldPath=$db->query("select Path from travelimage WHERE ImageID={$imageID}");
-    $oldPath=$rsOldPath->fetch_assoc()['Path'];
-    deletePic($oldPath);
+    if(isset($path)) {
+        global $path;
+        $rsOldPath = $db->query("select Path from travelimage WHERE ImageID={$imageID}");
+        $oldPath = $rsOldPath->fetch_assoc()['Path'];
+        deletePic($oldPath);
 
-    $db->query("update travelimage
+        $db->query("update travelimage
                  set Path= '{$path}'
                  WHERE ImageID='{$imageID}'");
 
-    if(!getCode($_POST['city'], $_POST['country'])){
-        rollback();
-        die("I am sorry. Is this place on the earth?");
-    }
-    $codes=getCode($_POST['city'], $_POST['country']);
+        if (!getCode($_POST['city'], $_POST['country'])) {
+            rollback();
+            die("I am sorry. Is this place on the earth?");
+        }
+        $codes = getCode($_POST['city'], $_POST['country']);
 
-    $db->query("update travelimagedetails
+        $db->query("update travelimagedetails
                 set Title='{$db->real_escape_string($_POST['photoName'])}', 
                     Description= '{$db->real_escape_string($_POST['description'])}',
                      Latitude={$_POST['latitude']}, 
@@ -93,9 +96,30 @@ function changeImage(){
                       CountryCodeISO= '{$codes['ISO']}'
                 WHERE ImageID={$imageID}");
 
-    if($db->error!=""){
-        rollback();
-        die($db->error);
+        if ($db->error != "") {
+            rollback();
+            die($db->error);
+        }
+    }else{
+        if (!getCode($_POST['city'], $_POST['country'])) {
+            rollback();
+            die("I am sorry. Is this place on the earth?");
+        }
+        $codes = getCode($_POST['city'], $_POST['country']);
+
+        $db->query("update travelimagedetails
+                set Title='{$db->real_escape_string($_POST['photoName'])}', 
+                    Description= '{$db->real_escape_string($_POST['description'])}',
+                     Latitude={$_POST['latitude']}, 
+                     Longitude={$_POST['longitude']}, 
+                     CityCode={$codes['GeoNameID']},
+                      CountryCodeISO= '{$codes['ISO']}'
+                WHERE ImageID={$imageID}");
+
+        if ($db->error != "") {
+            rollback();
+            die($db->error);
+        }
     }
 }
 
@@ -227,15 +251,18 @@ function deletePic($oldPath){
 }
 
 function rollback(){
-    global $db, $path;
+    global $db;
 
-    unlink("../img/travel-images/large/" . $path);
-    unlink("../img/travel-images/square-medium/" . $path);
-    unlink("../img/travel-images/square-small/" . $path);
-    unlink("../img/travel-images/square-tiny/" . $path);
-    unlink("../img/travel-images/medium/" . $path);
-    unlink("../img/travel-images/small/" . $path);
-    unlink("../img/travel-images/thumb/" . $path);
+    if(isset($path)) {
+        global $path;
+        unlink("../img/travel-images/large/" . $path);
+        unlink("../img/travel-images/square-medium/" . $path);
+        unlink("../img/travel-images/square-small/" . $path);
+        unlink("../img/travel-images/square-tiny/" . $path);
+        unlink("../img/travel-images/medium/" . $path);
+        unlink("../img/travel-images/small/" . $path);
+        unlink("../img/travel-images/thumb/" . $path);
+    }
 
     $db->query("rollback");
 }
